+++ src/main/java/app/controller/MultiplayerController.java
@@ -1,0 +1,12 @@
+package app.controller;
+
+import javafx.stage.Stage;
+
+class MultiplayerController extends Controller {
+
+  private static final String FXMLPATH = "view/multiplayer.fxml";
+
+  public MultiplayerController(Stage stage) {
+    super(stage, FXMLPATH);
+  }
+}